refactor(test): move image utils for tests in a separate module
