+++ pkg/test/common/utils.go
@@ -1,10 +1,15 @@
 package common
 
 import (
+	"errors"
 	"net/url"
+	"os"
+	"path/filepath"
 
 	"gopkg.in/resty.v1"
 )
+
+var ErrNoGoModFileFound = errors.New("test: no go.mod file found in parent directories")
 
 func Location(baseURL string, resp *resty.Response) string {
 	// For some API responses, the Location header is set and is supposed to
@@ -22,4 +27,26 @@ func Location(baseURL string, resp *resty.Response) string {
 	path := uloc.Path
 
 	return baseURL + path
+}
+
+func GetProjectRootDir() (string, error) {
+	workDir, err := os.Getwd()
+	if err != nil {
+		return "", err
+	}
+
+	for {
+		goModPath := filepath.Join(workDir, "go.mod")
+
+		_, err := os.Stat(goModPath)
+		if err == nil {
+			return workDir, nil
+		}
+
+		if workDir == filepath.Dir(workDir) {
+			return "", ErrNoGoModFileFound
+		}
+
+		workDir = filepath.Dir(workDir)
+	}
 }
